Address review finding: drop redundant long casts in port assertions
Review finding:

.github/instructions/java-tests.instructions.md says not to introduce redundant (long) casts in equalTo(longKey(...), value) when value is already an int, because the equalTo(AttributeKey<Long>, int) overload exists. ServerExtension.httpPort() returns int, and the three new assertions written by this PR cast it: `equalTo(SERVER_PORT, (long) server.httpPort())` here, and `equalTo(SERVER_PORT, (long) serverWithoutGreeter.httpPort())` twice in unknownService. The two pre-existing SERVER_PORT assertions in this same file, at lines 125 and 173, are written without the cast, so the new lines also read inconsistently with the code around them. Fix: drop the (long) casts from the three new SERVER_PORT assertions in this file.

Analysis: httpPort() returns int and equalTo has an AttributeKey<Long> overload that takes an int, so the cast changes nothing except how the line reads. Removing it matches both the documented test convention and the assertions already in this file. The (long) casts inside the nullable RPC_GRPC_STATUS_CODE ternaries stay, because there the cast picks the boxed type the conditional expression needs.

Upsides: The three new assertions match the repository's test convention and the surrounding assertions in the same file.

Downsides: No material downside identified.

+++ instrumentation/armeria/armeria-grpc-1.14/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/armeria/grpc/v1_14/ArmeriaGrpcTest.java
@@ -146,7 +146,7 @@ class ArmeriaGrpcTest {
                             equalTo(URL_PATH, "/example.Greeter/SayHello"),
                             equalTo(URL_SCHEME, "http"),
                             equalTo(SERVER_ADDRESS, "127.0.0.1"),
-                            equalTo(SERVER_PORT, (long) server.httpPort()),
+                            equalTo(SERVER_PORT, server.httpPort()),
                             equalTo(CLIENT_ADDRESS, "127.0.0.1"),
                             equalTo(NETWORK_PEER_ADDRESS, "127.0.0.1"),
                             satisfies(NETWORK_PEER_PORT, port -> port.isInstanceOf(Long.class)),
@@ -224,7 +224,7 @@ class ArmeriaGrpcTest {
                                 RPC_RESPONSE_STATUS_CODE,
                                 emitStableRpcSemconv() ? Status.Code.UNIMPLEMENTED.name() : null),
                             equalTo(SERVER_ADDRESS, "127.0.0.1"),
-                            equalTo(SERVER_PORT, (long) serverWithoutGreeter.httpPort()))
+                            equalTo(SERVER_PORT, serverWithoutGreeter.httpPort()))
                         .hasEventsSatisfyingExactly(
                             event ->
                                 event
@@ -242,7 +242,7 @@ class ArmeriaGrpcTest {
                             equalTo(URL_PATH, "/example.Greeter/SayHello"),
                             equalTo(URL_SCHEME, "http"),
                             equalTo(SERVER_ADDRESS, "127.0.0.1"),
-                            equalTo(SERVER_PORT, (long) serverWithoutGreeter.httpPort()),
+                            equalTo(SERVER_PORT, serverWithoutGreeter.httpPort()),
                             equalTo(CLIENT_ADDRESS, "127.0.0.1"),
                             equalTo(NETWORK_PEER_ADDRESS, "127.0.0.1"),
                             satisfies(NETWORK_PEER_PORT, port -> port.isInstanceOf(Long.class)),
